**离线混音器（`AudioMixer` + `AudioMixerScene`）多声道支持**：
1. **`AudioMixerTypes.h`** — 新增共享的 `ParseAudioFormatInfo()`，统一解析所有 OpenAL 布局（MONO/STEREO/QUAD/REAR/51CHN/61CHN/71CHN × 8/16/32-bit + 浮点）。

2. **`AudioMixer.cpp`** — 全链路按「帧 = 采样数 ÷ 声道数」重写：`ConvertToFloat`/`ConvertFromFloat` 增加 int32 分支、`ApplyPitchShift` 改为逐声道插值、`Mix` 改为帧/声道感知（`outputFrameCount = loopLength×sampleRate`，`outputSampleCount = frames×channels`）。

3. **`AudioMixerScene.cpp`** — 滤波器（低通/高通/带通/混响）改为逐声道独立状态；`ConvertFloatToOutput` 支持 int32；`GenerateScene` 用 `bytesPerFrame = bytesPerSample×channels` 计算字节，并新增 32-bit 整数混音分支。

+++ src/AudioResampler.cpp
@@ -120,7 +120,7 @@ namespace hgl::audio
                     float sample = input[i];
                     if(sample > 1.0f) sample = 1.0f;
                     if(sample < -1.0f) sample = -1.0f;
-                    dst[i] = static_cast<int32_t>(sample * 2147483647.0f);
+                    dst[i] = static_cast<int32_t>(static_cast<double>(sample) * 2147483647.0);
                 }
                 *output = data;
                 return;
